requirements.txt
streamlit>=1.42.0
openai>=1.65.0
pypdf>=5.3.0
python-docx>=1.1.2
Pillow>=11.1.0
pytesseract>=0.3.13
PyYAML>=6.0.2
reportlab>=4.2.5
Authlib>=1.3.2
